DAOS-14443 pool: Fix pool create performance
The "nominated" optimization for pool create operations in ds_rsvc has
become less effective because of certain recent timing changes. This
patch replaces the "nominated" optimization with new one that piggybacks
on the first POOL_CREATE sent to a pool service.

In theory, the absolute safety of both optimizations relies on a raft
update (yet to be made).

This patch also changes rsvc_client to begin each leader search from a
random replica, instead of a fixed one. When a lot of pool services are
created on the same set of ranks, this change distributes the pool
service leader more evenly.

Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true